+++ src/Deployer/Task/SystemTasks.php
@@ -16,6 +16,8 @@ class SystemTasks extends TaskAbstract
 {
     const TASK_PHP_FPM_RESTART = 'sys:php-fpm:restart';
     const TASK_NGINX_RESTART = 'sys:nginx:restart';
+    const TASK_CRON_STOP = 'sys:cron:stop';
+    const TASK_CRON_START = 'sys:cron:start';
 
     public static function register()
     {
@@ -28,6 +30,14 @@ class SystemTasks extends TaskAbstract
             self::TASK_NGINX_RESTART, 'Restart nginx service',
             function () { SystemTasks::restartNginx(); },
             ['staging', 'production']
+        );
+        Deployer::task(
+            self::TASK_CRON_STOP, 'Stop cron service',
+            function () { SystemTasks::stopCron(); }
+        );
+        Deployer::task(
+            self::TASK_CRON_START, 'Start cron service',
+            function () { SystemTasks::startCron(); }
         );
     }
 
@@ -47,6 +57,24 @@ class SystemTasks extends TaskAbstract
     {
         $service = \Deployer\get('nginx_service');
         \Deployer\run("sudo service $service restart");
+    }
+
+    /**
+     * Stop cron-daemon using cron_service variable
+     */
+    public static function stopCron()
+    {
+        $service = \Deployer\get('cron_service');
+        \Deployer\run("sudo service $service stop");
+    }
+
+    /**
+     * Start cron-daemon using cron_service variable
+     */
+    public static function startCron()
+    {
+        $service = \Deployer\get('cron_service');
+        \Deployer\run("sudo service $service start");
     }
 
 }